Created Change Username/Password for Account Recovery
I have created the UI layout for the page where users can change their username/password for account recovery. I decided that changing username/password can be on the same page for simplicity. I made the forms to change username/password as components because it made the main page longer than it needed to be. Still have yet to add functionality.