Can now look at chat logs of different people in contacts list
I have added functionality that allows a user to click on each member in their contacts list and see the messages sent between them and that person chosen in particular. Also makes sure that the message log profile pictures and names displayed take that into account as well. Took me till 3 to do this because I had to learn about use states for the first time in like 2 years and I'm kind of an idiot sooooo