Fixed Errors in Direct Messaging Page
I fixed the errors in the direct messaging page. Turns out, setting '' as the default is NOT a good idea, and it only showed that it was fine on my end this morning because of the way useState works. I made it so that the default should be whoever is at the top of their contacts list. Still doesn't handle any situation for if there is no one in the contacts list, but that should be easy to implement.